Changes to replace generation with alt_generation
