+++ electricitylci/combinator.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 import pandas as pd
 from electricitylci.globals import output_dir, data_dir
-import electricitylci.alt_generation as altg
+import electricitylci.generation as gen
 import electricitylci.import_impacts as import_impacts
 from electricitylci.model_config import (
         eia_gen_year, 
@@ -240,8 +240,8 @@ def concat_clean_upstream_and_plant(pl_df, up_df):
     Parameters
     ----------
     pl_df : dataframe
-        The generator dataframe, generated by electricitylci.generation or
-        .alt_generation
+        The generator dataframe, generated by electricitylci.generation
+        
     up_df : dataframe
         The combined upstream dataframe.
 
@@ -422,7 +422,7 @@ if __name__ == "__main__":
     upstream_df = concat_map_upstream_databases(
         coal_df, ng_df, petro_df, geo_df, nuke_df, solar_df, wind_df
     )
-    plant_df = altg.create_generation_process_df()
+    plant_df = gen.create_generation_process_df()
     plant_df["stage_code"] = "Power plant"
     print(plant_df.columns)
     print(upstream_df.columns)
